server/requirements.txt
aiohttp==3.8.4
keyboard==0.13.5
pycryptodome==3.23.0
python-dotenv==1.1.0
python_bcrypt==0.3.2
websockets==15.0
